requirements.txt
fairseq
datasets
dataclasses
torchmetrics
numba
librosa
deepspeed
mpi4py
optuna
peft
bitsandbytes
